Create verndors_compare.js
the selection of the best vendor should be based on some criteria of the vendor. The most important criterion for this is the price, the lower, the better.
If the prices are the same, we look at other criteria, the first is the shipping time (the shorter, the better), after that the rating (the higher and more the better) and availability (the more, the better) are to be ranked.
The availability is always greater than 0. The seller with the best result will be displayed.